重置 忽略列表
推进版本

+++ AorFramework/Unity5/Assets/Resources/shader/Custom/Light/Diffuse - Toon -TangentCheck.shader
@@ -37,7 +37,7 @@ Shader "Custom/Light/Diffuse - Toon - TangentCheck" {
 
 				float _Outline;
 				float4	_OutlineColor;
-				int _HideFace;
+				float _HideFace;
 				sampler2D _ToonShade;
 				sampler2D _MaskTex;
 				sampler2D _NormalTex;
